py/pyproject.toml
[project]
name = "image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
